third_party/blink/web_tests/platform/generic/fast/backgrounds/multiple-backgrounds-initial-values-expected.txt

Test that background position values do not leak between layers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS bgPosLayers[0].search("-10px") is >= 0
PASS bgPosLayers[0].search("50px") is >= 0
PASS bgPosLayers[1].search("-10px") is -1
PASS bgPosLayers[1].search("50px") is -1
PASS successfullyParsed is true

TEST COMPLETE
